Finished making events a protected endpoint

+++ api/routers/events.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, Response
+from fastapi import APIRouter, Depends, Response, status
 from typing import List, Optional, Union
 from queries.events import (
     Error,
@@ -6,53 +6,70 @@ from queries.events import (
     EventQueries,
     EventOut,
 )
+from queries.users import UserOut
+from authenticator import authenticator
 
 router = APIRouter()
 
+def get_current_user(user: UserOut = Depends(authenticator.get_current_account_data)):
+    return user
 
-@router.post("/api/events", response_model=Union[EventOut, Error])
+@router.post("/api/events/", response_model=Union[EventOut, Error], status_code=status.HTTP_201_CREATED)
 def create_event(
     event: EventIn,
     response: Response,
+    current_user: UserOut = Depends(get_current_user),
     query: EventQueries = Depends(),
 ):
-    try:
-        result = query.create(event)
-        response.status_code = 200
-        return result
-    except Exception as e:
-        response.status_code = 400
-        return {"Error": str(e)}
+    if current_user:
+        try:
+            result = query.create(event)
+            response.status_code = 201
+            return result
+        except Exception as e:
+            response.status_code = 400
+            return {"Error": str(e)}
+    else:
+        return "You need to log in to create an event"
 
 @router.put("/api/events/{event_id}", response_model=Union[EventOut, Error])
 def update_event(
     event: EventIn,
     event_id: int,
+    current_user: UserOut = Depends(get_current_user),
     query: EventQueries = Depends(),
 ) -> Union[Error, EventOut]:
-    return query.update(event_id, event)
+    if current_user:
+        return query.update(event_id, event)
 
-@router.get("/api/events/", response_model = Union[List[EventOut], Error])
+@router.get("/api/events/", response_model=Union[List[EventOut], Error])
 def get_all_events(
+        current_user: UserOut = Depends(get_current_user),
         query: EventQueries = Depends()
 ):
-    return query.get_all_events()
+    if current_user:
+        return query.get_all_events()
+    else:
+        return "You need to log in to view this"
 
-
-@router.delete("/api/events/{event_id}", response_model =bool )
+@router.delete("/api/events/{event_id}", response_model=bool)
 def delete_event(
     event_id: int,
+    current_user: UserOut = Depends(get_current_user),
     query: EventQueries = Depends(),
 ) -> bool:
-    return query.delete_event(event_id)
+    if current_user:
+        return query.delete_event(event_id)
 
-@router.get("/api/events/{event_id}", response_model = Optional[EventOut])
+@router.get("/api/events/{event_id}", response_model=Optional[EventOut])
 def get_event(
     event_id: int,
     response: Response,
+    current_user: UserOut = Depends(get_current_user),
     query: EventQueries = Depends(),
 ) -> EventOut:
-    event = query.get_event(event_id)
-    if event is None:
-        response.status_code = 404
-    return event
+    if current_user:
+        event = query.get_event(event_id)
+        if event is None:
+            response.status_code = 404
+        return event
